+++ packages/react-core/src/components/Label/__tests__/Label.test.tsx
@@ -58,3 +58,8 @@ test('label with additional class name and props', () => {
   );
   expect(view).toMatchSnapshot();
 });
+
+test('label with truncation', () => {
+  const view = shallow(<Label isTruncated>Something very very very very very long that should be truncated</Label>);
+  expect(view).toMatchSnapshot();
+});