+++ packages/react-core/src/components/TreeView/index.ts
@@ -1,0 +1,4 @@
+export * from './TreeView';
+export * from './TreeViewList';
+export * from './TreeViewListItem';
+export * from './TreeViewSearch';