+++ packages/react-integration/demo-app-ts/src/components/demos/NavDemo/NavDemo.tsx
@@ -25,6 +25,7 @@ export class NavDemo extends Component {
     expandableActiveGroup: 'grp-1',
     expandableActiveItem: 'itm-1-1',
     horizontalActiveItem: 0,
+    tertiaryActiveItem: 0,
     expandableClickedGroup: '',
     expandableClickedItem: ''
   };
@@ -263,6 +264,44 @@ export class NavDemo extends Component {
     );
   }
 
+  onTertiarySelect = (result: SelectedItem) => {
+    this.setState({ tertiaryActiveItem: result.itemId });
+  };
+
+  renderTertiaryNav() {
+    const { tertiaryActiveItem } = this.state;
+
+    return (
+      <StackItem>
+        <Title headingLevel="h2" size="2xl">
+          Tertiary Nav
+        </Title>
+        <div
+          style={{
+            borderStyle: 'solid',
+            borderWidth: '0.2px',
+            borderColor: '#292e34',
+            padding: '1rem'
+          }}
+        >
+          <Nav onSelect={this.onTertiarySelect} id="nav-primary-tertiary" variant="tertiary">
+            <NavList>
+              <NavItem id="tertiary-link1" preventDefault itemId={0} isActive={tertiaryActiveItem === 0}>
+                Item 1
+              </NavItem>
+              <NavItem id="tertiary-link2" preventDefault itemId={1} isActive={tertiaryActiveItem === 1}>
+                Item 2
+              </NavItem>
+              <NavItem id="tertiary-link3" preventDefault itemId={2} isActive={tertiaryActiveItem === 2}>
+                Item 3
+              </NavItem>
+            </NavList>
+          </Nav>
+        </div>
+      </StackItem>
+    );
+  }
+
   render() {
     // Nav onToggle and onSelect should be optional
     // https://github.com/patternfly/patternfly-react/issues/1234
@@ -271,6 +310,7 @@ export class NavDemo extends Component {
         {this.renderDefaultNav()}
         {this.renderExpandableNav()}
         {this.renderHorizontalNav()}
+        {this.renderTertiaryNav()}
       </Stack>
     );
   }