+++ packages/react-topology/src/layouts/LayoutNode.ts
@@ -1,7 +1,9 @@
+import * as d3 from 'd3';
 import { Node, NodeStyle } from '../types';
 import { Rect } from '../geom';
 import { LayoutGroup } from './LayoutGroup';
-export class LayoutNode {
+
+export class LayoutNode implements d3.SimulationNodeDatum {
   protected readonly node: Node;
   protected xx?: number;
   protected yy?: number;
@@ -58,6 +60,9 @@ export class LayoutNode {
         .clone()
         .setCenter(x, y)
     );
+  }
+  setFixed(fixed: boolean): void {
+    this.isFixed = fixed;
   }
   get nodeBounds(): Rect {
     const { padding } = this.node.getStyle<NodeStyle>();