+++ packages/react-docs/CHANGELOG.md
@@ -3,6 +3,113 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.8.26](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.8.25...@patternfly/react-docs@5.8.26) (2020-09-02)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.8.25](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.8.24...@patternfly/react-docs@5.8.25) (2020-09-01)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.8.24](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.8.23...@patternfly/react-docs@5.8.24) (2020-08-28)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.8.23](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.8.22...@patternfly/react-docs@5.8.23) (2020-08-27)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.8.22](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.8.21...@patternfly/react-docs@5.8.22) (2020-08-26)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.8.21](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.8.20...@patternfly/react-docs@5.8.21) (2020-08-25)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.8.20](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.8.19...@patternfly/react-docs@5.8.20) (2020-08-25)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.8.19](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.8.18...@patternfly/react-docs@5.8.19) (2020-08-25)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.8.18](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.8.17...@patternfly/react-docs@5.8.18) (2020-08-24)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.8.17](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.8.16...@patternfly/react-docs@5.8.17) (2020-08-24)
+
+
+### Bug Fixes
+
+* **docs:** use ids and h3s ([#4720](https://github.com/patternfly/patternfly-react/issues/4720)) ([a3f684f](https://github.com/patternfly/patternfly-react/commit/a3f684fa23cfce3d8e79b86206773292790cec5c))
+
+
+
+
+
+## [5.8.16](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.8.15...@patternfly/react-docs@5.8.16) (2020-08-24)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.8.15](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.8.14...@patternfly/react-docs@5.8.15) (2020-08-24)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.8.14](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.8.13...@patternfly/react-docs@5.8.14) (2020-08-24)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
 ## [5.8.13](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.8.12...@patternfly/react-docs@5.8.13) (2020-08-24)
 
 **Note:** Version bump only for package @patternfly/react-docs